fix(levm)!: fix jumpdests for large initcodes (#5254)
> [!CAUTION]
> This changes the serialization of `Code`, and thus requires
re-syncing.

**Motivation**

While executing the state tests in debug mode, [a debug
assert](https://github.com/lambdaclass/ethrex/blob/45a9871498bd155867c8e9a9bf86e73fc634841a/crates/common/types/account.rs#L44)
was triggered.

The calculation of valid jumpdests assumes the code will be less than
2^16 bytes in length, however this isn't true before Shanghai.

**Description**

This PR adds support for up to 2^32 bytes of initcode.

---------

+++ crates/common/types/account.rs
@@ -24,8 +24,11 @@ use crate::{
 pub struct Code {
     pub hash: H256,
     pub bytecode: Bytes,
-    // TODO: Consider using Arc<[u16]> (needs to enable serde rc feature)
-    pub jump_targets: Vec<u16>,
+    // TODO: Consider using Arc<[u32]> (needs to enable serde rc feature)
+    // The valid addresses are 32-bit because, despite EIP-3860 restricting initcode size,
+    // this does not apply to previous forks. This is tested in the EEST tests, which would
+    // panic in debug mode.
+    pub jump_targets: Vec<u32>,
 }
 
 impl Code {
@@ -40,8 +43,8 @@ impl Code {
         }
     }
 
-    fn compute_jump_targets(code: &[u8]) -> Vec<u16> {
-        debug_assert!(code.len() <= u16::MAX as usize);
+    fn compute_jump_targets(code: &[u8]) -> Vec<u32> {
+        debug_assert!(code.len() <= u32::MAX as usize);
         let mut targets = Vec::new();
         let mut i = 0;
         while i < code.len() {
@@ -49,7 +52,7 @@ impl Code {
             match code[i] {
                 // OP_JUMPDEST
                 0x5B => {
-                    targets.push(i as u16);
+                    targets.push(i as u32);
                 }
                 // OP_PUSH1..32
                 c @ 0x60..0x80 => {
